Add resource bindings and accessors to effect.
Implements resources bindings and accessors in the Effect class enabling
the user to easily assign values to their compiled and linked shader
programs. Currently does not support binding uniforms to a specific
index but rather retrieves and caches the automatically assigned uniform
index the first time the uniform is written to.

Added debug helper functions to the Effect class that allow the user to
print out a list of currently active uniforms and attributes on their
shader program.